Make system image selection less flaky
The system image table is the table of system images
that a user can download or use to create an AVD. This
table appears when creating an AVD. The table is populated
with information about installed system images and system
images that are available for download. However, the system
images that are available for download appear later after
the table is already populated with installed system images.
This is a potential spot for a TOCTOU bug because the test
can sometimes check that the table has the requested image
in a certain row before the downloadable system images
are shown. When the downloadable system images appear,
the test ends up clicking on the wrong row.

This change fixes this issue by having the test check
which row it selected before moving on to the next step
of creating the AVD.

Test: InstantRunTest.coldSwap
Change-Id: I9dde5de593d4bf9429832556dca42b93613a1096
Bug: 71608080