Add Google Maven repository when updating Android plugin
If an older version of a project is updated to use a newer version of
Gradle and AGP, this new plugin can be found in the Maven repository, if
the repository is not found, then calling gradle from the command line
will fail.

This CL adds Google Mavel repository to buildscript when updating plugin
version.

Test: AndroidPluginVersionUpdaterIntegrationTest#testUpdatePluginVersion
now checks that the reporitory is present after the update.

Bug: 69977310
Change-Id: Ida24e92bb234b019055475061da62140850c083d