Add horizontal & vertical dividers to palette.
The dividers are added to the bottom of the
Widget group.
The XML generated will be a <view> where the
width of the divider is set to 1dp and the
length is set to "match_parent".
The "match_parent" value will be converted to
a constant when dropped into a Constraint
Layout.

In this CL the icon on the component tree is
also updated to show a divider icon.

Bug: b/37137000
Test: Added unit tests.
Change-Id: I2b5825d740eded27c70830f523717c08c8ee7a4e