Fix missing synchronization around refreshing menu items
TemplateManager has a pretty weird lifecycle. It is supposed to be an
application wide thing, but also refreshes the menu across the
application after each gradle sync in any open project.

The code that refreshes the menu after a sync isn't synchronized at all.
If you happen to have two syncs complete approximately at the same time,
both will end up attempting to refresh the menu, and you'll see the same
menu items twice.

This CL fixes the issue by moving the unsynchronized code to be under
the same lock as the rest of the refresh menu code.

Test: manual, launch two projects at the same time
Change-Id: I2f03bd93b53d1cf666c1fcaa0016745c3e2cb8eb
Fixes: 71853976