Disable UninstallIotLauncherAppsTask for 3.1
In the upcoming DP7 release of Android Things in February, we plan to
change the auto-launch mechanism of apps. Since we don't want Android
Studio 3.1 to ship with a functionality that may break in February, we
are disabling the uninstall feature with a feature flag. We will
reintroduce it in Android Studio 3.2 after Android Things DP7 launches.

Test: Updated InstantRunBuildAnalyzerTest
Bug: 69863386
Change-Id: Ie54e748f3ad4d47c70b8dbe65ff16e9dd9bb1624