Fix modality state in SetPreference
We were using ModalityState.any() to run a write action, which is
now forbidden. Updating to use the correct modality state (that
of the wizard).

Fixes: 71697590

Test: The first run wizard doesn't have many tests, currently.
My understanding is that it's on its way out (eventually? Soon?)
so I don't think it makes sense to invest too much time in doing
tests for it now.

Change-Id: Ic290a0c32a5bdb85a03e7e439fd79bcd33c386de