Initialize package name lazily in AndroidResourceFilesListener
AndroidResourceFilesListener caches the current value of package name
from the manifest. The value is used to trigger some source generation.
All of this is applicable only for JPS projects, and even there it is
purely an optimization since it is only used in the context of a VFS
change event. This CL deletes this optimization, primarily to remove the
dependency from the facet to AndroidResourceFilesListener.

Bug: 71553525
Test: existing tests pass
Change-Id: I9f9a345e25459535ac24fa928e37bf0a9fabbfcb