Not dying if nav dependency isn't added + ui tests
Instead we just won't continue initializing the editor, and we'll show
the error screen. This should fix the problem with
NavNlEditorTest.testAddDependency.

There was a problem with NavNlEditorTest where sometimes we would
try to start interacting with the DesignSurface before it was showing.
Added appropriate waits for that.

We were registering PropertiesManager as a disposable child of the
project, but all ToolContents are explicitly disposed by their
containing AttachedToolWindow.

NonEditableEditor and TextEditor weren't getting cleaned up properly
when the project was closed (rather than the editor being closed)
because they weren't initialized with a project.

NavPropertiesPanel wasn't being registered as a disposable child of
anything.

Fixes: 71804935, 69000941, 70305086

Test: fixes NavNlEditorTest

Test-Strategy: ui.test.groups=DEFAULT

Change-Id: I8f54d3bf940beb169c71ddafcf17845e1f7fd3da