Reduce the scope of the PsiTreeChangeListener in GradleFiles.
This change restricts the PsiTreeChangeListener to only be active when a
Gradle build file or external build file has been selected in the
editor. We do this be adding a FileEditorManagerListener then adding or
removing the PsiTreeChangeListener based on the newly selected file.

The goal of this is to not have the PsiTreeChangeListener active all the
time since these events are triggered a lot. Even though this change
adds the FileEditorMnaagerListener which is always active, the number of
events it has to deal with is significantly less then the
PsiTreeChangeListener (only when a new file is selected in the editor).

Change-Id: I60ab99a0c0565e052bf04e1efcc454c4db1fc6e1
Fixes: 71816380
Test: Unit test added.