Re-enable test cases that were temporarily disabled
Tests had originally been disabled because problems had been
observed when running the tests locally on certain versions
of Gradle.

The earlier approach had been to count the number of paths
and configurations returned based on different criteria.
However, Gradle will often return paths to potential file
locations that aren't actually present on the file system.
Either due to bugs or intentional changes in the gradle
plugin, different versions of gradle return different
hypothetical paths. Since the paths don't actually exist
and Gradle is allowed to return extra paths to things
that don't exist the test should be able to deal with such
changes without failing.

The new approach is to only check for locations that were
actually created on disk as part of the test fixture, and
ignore extra no-op configs or file locations.

Bug: 71553839
Test: GradleModelConverterTest
Change-Id: Idb4f1cb347873cf9368528f4e1c58b5e5b5991da