Stabilize ConstraintLayoutTests.
In EditorFixture.getLayoutEditor, the layout editor can still be
initializing. The method should wait for idleness
before continuing.

In NlEditorFixture.dragComponentToSurface, the component can
still move slightly after being dropped to the surface. This can
throw off the next click of the test, so this method should also
wait for idleness before continuing.

Test: ConstraintLayoutTest
Change-Id: I3f28d5a7f2ee504d36a00fbcf292bf0e6f29b16d