[AIA] Create Provision Task only for aia projects
When adding a Provision before run task to an Android run configuration,
now we check if the project is an Instant App project (has at least one
module of type PROJECT_TYPE_INSTANTAPP) and only in this case we add it.
This will make users less confused when not using instant apps at all.

Test: ProvisionBeforeRunTaskIntegTest
Bug: 71599222
Change-Id: Ief05fd23b847a25af3d5f77d7ff61ec1b339872d