Do not show tint if not present for ProgressBar
The attributes "progressTint" and "indeterminateTint" only exist for
ProgressBar in API >= 21. If in a project with minApi < 21, opening the
attributes table of the layout editor for a progress bar will trigger an
error as those attributes will not exist.
So if the attributes cannot be found, do not try to display an editor
for them.

Test: ConstraintLayoutTest.addAllLayoutItemsFromToolbar
Change-Id: I1acda4e8b981520ed559d181815977104bb0f2d9
(cherry picked from commit 73b5ec1e98bc7d60f8c6363d35ba964b82ebd13a)