[Gradle Build Model] Add support for editing map properties.
This CL enables the creation, modification and deletion of map
properties. In order to do this we had to fix how GradleDslReferences
are created, how maps are created and make sure that any non singular
value that is set in a map is preceded by a comma.

We also fix an issue with deleting element whereby GradleDslElements
had PsiElements that were invalid, this caused us to not be able to
recreate them if it was required.

We do not currently support parsing OR setting maps within maps.

Bug: 67519210
Test: Unit tests added.
Change-Id: Ie45d4c8d1a4d9a3b79b024926765d9f13df24ad5