Integrate transplanted testGuiFramework
JetBrains forked our UI test framework code about 2 years ago, creating
testGuiFramework. The structural modifications therein allow writing
tests that restart the IDE, as well as tests that run on the release.

This change modifies the files copied over in the previous commit, in an
attempt to integrate that portion of testGuiFramework into our
uitest-framework module.

Under this framework, when a test is to be run, the "server" will send a
message to a running "client" instance of Studio (that includes the
uitest-framework module) instructing it to run the specified test. The
client will respond with the test results when done.

The server can launch a new client either from the compiled classes in
the tools/idea/out directory, or by running studio.sh on a release
build. In either case, the client runs in a separate JVM and
communicates with the server via a socket. This separation allows the
server to persist across client IDE restarts, so multi-part tests can be
written where the IDE restarts between each part. This is useful for
testing the patcher, for example.

Initially, only one test, RestartIdeTest, will run under the new framework.
The current setup for running existing UI tests with GuiTestRunner should
still work.

Test: RestartIdeTest
Test-Strategy: ui.test.groups=TEST_FRAMEWORK,THEME
Bug: 74114630
Change-Id: I66173a30328126afd6be0acb192999ba05963b08