[AVD Mgr] Change a "Graphics" option that we give to the Emulator
When the user creates an AVD with the Graphics:Software option,
we used to pass the setting to the Emulator with
"hw.gpu.mode=guest". The Emulator team wants a different default
software renderer. Let's tell them "hw.gpu.mode=software" and have
them pick the specific renderer that they think is best.

Unfortunately, we still need to use the old setting with old emulators.

Fixes: 72090948

Test: Created GpuModeTest.kt

Change-Id: I1b6879fc6e8766cfc25956d2daea2bf1208082f6