Delete eager creation of ResourceRepositories
ResourceRepositories is created eagerly in AndroidFacet, but the
initialization doesn't really do much other than creating an object and
associating it with the facet. This behavior was introduced as part of a
cleanup in ag/I5c9edb03e107a210. This CL removes this eager
initialization getting us a little closer to removing unnecessary
dependencies from the facet.

Test: existing tests pass
Fixes: 71553525
Change-Id: I91285f4bab7a29b2ad2cf8f7d03513f52bc40695