[Gradle Build Model] Support recursive maps
This change supports creating, setting and removing maps within maps
using the GradlePropertyModel.

In order to do this we had to extend GradleDslExpressionMap to provide
whether it is used in a literal expression map such as "prop2 =
[key : 'value']" or used as a command/method arguments i.e apply plugin:
'value'.

We use this information to provide an alternative way to create maps
when they are literal maps and are DERIVED.

Bug: 72099838
Test: Unit tests added.
Change-Id: Ief6aac4b24e85b63333e97be6a35b42032f252a0