Add log statement to point out null values.
It's possible the output parser may be null values from
the deserialization of the JSON output from Gradle build
output. This can lead to a null pointer exception in
bug 72164080.

This change adds a log statement to determine whether
or not this is occurring on the CI system, since this
issue is difficult to reproduce locally.

Bug: 72164080
Test: NewKotlinProjectTest
Change-Id: I5a2c10ef89327993db5026b4f31522edcf143544