Remove the dependency of AndroidFacetConfiguration on DataBindingUtil
Previously, AndroidFacetConfiguration was hardcoded to call
DataBindingUtil in order to notify ModuleDataBinding whenever
the model changed. This dependency is unnecessary because there
is already a message bus that fires events whenever the model
changes.

This change adopts a message bus subscriber in ModuleDataBinding
and removes the DataBindingUtil.refreshDataBindingStatus method
that previously performed these updates.

Bug: 71553325
Test: Used a breakpoint in ModuleDataBinding.setEnabled to confirm that
the data binding state gets toggled whenever data binding is enabled in
the project

Change-Id: I6ff84441714b30dcb367c3756270aceb74783b87