Fix sporadic failure of GradleTaskRunnerTest
I guess the cause is timeout: invocation of Gradle with a custom
BuildAction was eventually lasting more than 1 minute.
However, this test was also keeping a reference to Project that was
already disposed in the end, because DefaultGradleTaskRunner keep a
reference to an instance of project and was not disposable.

This CL increases timeout, fixes dealing with Exception in the second
thread in the test and registers a Disposable in DefaultGradleTaskRunner
to remove the reference to Project.

Test: GradleTaskRunnerTest.testBuildActionRunner
Fix bug 72262273

Change-Id: I631f36916a407135f62816d6abb9dda21b7b830b