Fixed ProfilerTimeline's incorrect initialization of duration
Before this change, the ProfilerTimeline is reset with a duration value
to account for the time between beginSession and the timeline being reset.
It turns out to be incorrect as that duration should already be accounted
for by the update loop. Because the duration value was extremely small before,
the issue had gone unnoticed. But the change to move attach-agent alongside
beginSession (which takes 1+ second) exposes the problem.

Bug: 67508564
Test: updated
Change-Id: I6c302067636bd7ada7661f3e5e7e02e3a25a7df8