[AVD Mgr] Fix emulator version that supports new GPU option
We added a new option for the "hw.gpu.mode" option in the config.ini.

This new option is only for use with newer emulator versions. The
cutoff version was incorrectly set at 27.0.5. The correct cutoff
is 27.1.6.

The previous change: ca7cd284c6948fc34504b7a139747c0d8a8603d0

Fixes: 72262742

Test: None (GpuModeTest still passes)

Change-Id: I022a00e24c1c836ffd782516c7a742b4829418fb