Add basic layout editor UI tests targeting Bazel project
Currently, NlEditorTest#basicLayoutEdit fails unexpectedly when
run via host Bazel and targeting a Bazel project. This failure
occurs because the test expects the design editor to be usable
immediately after the initial sync, when in reality the editor
waits for indexing to complete before rendering.

This CL introduces NlEditorFixture#waitForSurfaceToLoad and
NlEditorFixture#canInteractWithSurface, which allow the
basicLayoutEdit test to ensure that the design editor is usable.
The waitForSurfaceToLoad method determines this based on the
visibility of the design surface itself, rather than the progress of
background tasks.

Additionally, this CL add tests to cover what happens when the
design editor can't render due to a failing sync when working
with a Bazel project. These tests are meant to prevent regressions
affecting the recent work to make layout editor initialization
build system independent (see ag/65213155).

Note that NlEditorTest#basicLayoutEdit still fails when run from Bazel,
but for unrelated reasons.

Bug: 71713967, 71722519
Test: new tests pass when run with Bazel
      + NlEditorTest#basicLayoutEdit no longer fails from timeout issue
        when run with Bazel.
Change-Id: Ib9fa446cd80412bed9dfedca0aab9cc7a4570a82