Updating StateChart to not depend on Enum generic type.
This change is required for CPU thread coloring. The threads used in the state chart
will be passed by thread name/id as such we cannot map every name/id to an enum value.
Instead we will map the name to a function that determines if this thread is part of
the current process. If it is it will be colored one way, otherwise it will be colored
a different way.

Test: Internal Structure Change Test shouldn't be affected.
Bug: 64693256
Change-Id: Ia3495ab2bf239632b565eedf40365839899d0cd7