Implement PathStringPool
PathStringPool is a utility for converting arbitrary PathString
objects (which may not exist on the filesystem) into temporary
File objects, VirtualFiles, or InputStreams. The files are deleted
when the converter is closed. This can be used for code which has one
of VirtualFile, Path, PathString, or File and needs to invoke a method
that accepts a different one of the above.

Bug: 67467865
Test: PathStringPoolTest
Change-Id: Icf4db209e0d918794b16d0b6d976089eaf3828f4