[Gradle Build Model] Rework GradlePropertiesDslElement.
This change reworks the internals of the GradlePropertiesDslElement to
allow for the storing of multiple GradleDslElements per property name.
This is in preparation for adding scoping and dependency information for
variables.

In order to do this we create a holder class ElementList which stores
an collection of ElementItems which attach extra state to the
GradleDslElement.

Fixes: 71703978
Test: Unit tests already exist
Change-Id: I4f6719a659457be818a8830e39016155ce324dad