Cache GuiTestProjectSystem in a rule in ui-tests.
Currently the GuiTestProjectSystem used during ui-tests is fetched from
extension point every-time it is needed. This logic should not be
handled by GuiTestRule and clutters the class.

This CL encloses the GuiTestProjectSystem in a rule called CurrentProjectSystem
that extends ExternalResource much like the RobotTestRule.  The rule caches
the GuiTestProjectSystem and simplifies GuiTestRule by moving the logic out.

Test: Manually tested ui-tests to verify the behaviour stays the same.
Change-Id: I4c3475e9e8f525f50fece80dbb04ba2df41e7376