Synchronize access to GuiTestSuiteState
In about ~2 months of running UI tests, there has been a single report
of a failure where getGuiTestSuiteState returned null when it is
annotated as @NotNull.

There were two possible issues with the code - one is that even though
the constructor used to set the UI testing mode, it did not actually
initialize the state. The more likely cause is a race condition between
the thread that enables ui testing mode vs the thread that retrieves the
state.

This CL fixes both possible issues.

Fixes: 72644561
Test: existing UI tests
Change-Id: Icfd4dcaee18a744b4922ba1eb0f21c5442644657