Remove one source of flakiness in createAndRunInstantApp
This test can flake if it opens the AVD manager window before the
project has finished setting up. When the project setup finishes,
the main IDE window can steal focus and appear on top of the AVD
manager dialog window. This causes the test to fail because it is
unable to find the manager dialog window.

Bug: 72666461
Test: InstantAppRunTest.createAndRunInstantApp
Change-Id: I7cea2197bed04885f9d46824e302150db51bccb4