Add value property to CommonComboBoxModel.
A ComboBox should be able to show and edit values that doesn't
exist in the dropdown list. Add support for this by extending
a CommonComboBoxModel to extend a CommonTextFieldModel.
That same model will be used for the text field that is created
by the UI to facilitate editing on the combo box value.

Test: Updated/extended.
Change-Id: I89a78f658188a2618db53a6d6f6259ddd14f930b