project already disposed in NlDesignSurface
The Layout Editor Surface should stop rendering when the current
project is closed. However, it is possible that it makes a call to
GradleProjectInfo.isBuildWithGradle when the project is still not
disposed, and because isBuildWithGradle runs inside a read lock, by the
the time isBuildWithGradle() is executed the project is disposed.
This CL checks for project disposed, when the read locks executes and
follows up on Change-Id Iadea42073445b61a414ad1e2c0aac5c805d8f4b6

Change-Id: Idd56095267294c4f9dce65624ba61eca4cc2ec62
Fixes: 70682962
Test: NewActivityTest.createActivityWithFragment