Make LayoutInspectorTest wait for more fine-grained events
LayoutInspectorTest.launchLayoutInspectorViaChooser was too flaky. The
flakiness was caused by the test not waiting for enough fine-grained
events to occur on the emulator.

There are two events that a test can currently wait for when launching
an application on the emulator:

1. Wait for a connection to the application process on the emulator.
2. Wait for the emulator to run and render an application activity.

The test was only waiting for the second event to occur within a span of
5 seconds. This is often not enough time for the emulator to finish
booting, launch the app, place the main activity into the run state, and
to draw the activity on the screen. The test should wait for the first
event before waiting on the second event.

Bug: 71361448
Fixes: 71361448
Test: LayoutInspectorTest.launchLayoutInspectorViaChooser
Change-Id: Ie331adf05f2a39086fe4ed1dcfb468e14a1a81f2