android studio AIA lib integration
Adds the ability to run Instant Apps using the upcoming Instant Apps SDK
Library distributed as part of the Instant Apps SDK. This is an
alternate launch path for Instant Apps. The library handles
provisioning, deploying, and launching the app.

Unfortunately test coverage isn't quite where I'd like it to be.
Specifically:
- Integration tests don't seem to be viable until we have the ability to
release the library (correct me if I'm wrong on this), and that's still
behind some approvals and eng cleanup. I have verified manually that it
works.
- There is no test class for AndroidLaunchTasksProvider and I'm not sure
if I should create one, or if trying to do so would be a rabbit hole, so
I'll the advice of my reviewer.

Change-Id: I38a82b9e8aeab90c6eff1ef9882dac1cf3ce4fae
Test: Tested manually. Has some unit tests, integration tests don't
appear to be possible yet because the SDK lib hasn't been released.
Fixes: 72222169