Fix 'UElement as PsiElement usage' inspection finding
There's a new inspection in IntelliJ 2018.1 meant to help eventually
disentangle the UElement class hierarchy from the Psi hierarchy
(UMethod no longer extends PsiMethod, etc).

UElement points to the actual "source" Psi (with the caveat that for
Kotlin we can't operate on it directly), and to a "Java-like" Psi that
allows basic operations to work in a language-agnostic way (and for
certain Kotlin constructs this may be null, because there's no
correspondent in Java).

Whenever a UElement is used as a PsiElement, we need to disambiguate
the use case.

Test: existing
Change-Id: Id36036bda811c7838c0bfeb73cf31f4fd7b4af6d