Adjust timeout flakiness in createAndRunInstantApp
InstantAppRunTest.createAndRunInstantApp does not have enough time
to wait for the emulator to start up. In the case of instant apps,
the device needs to be provisioned, and device provisioning takes much
longer compared to booting an emulator normally.

This change adjusts the timeout for createAndRunInstantApp by waiting
up to 3 minutes for the emulator to be booted and provisioned.

Test: InstantAppRunTest.createAndRunInstantApp
Bug: 72815713
Change-Id: I9ecd3af94725a5a91a006fb209f42705c5e97f04
Fixes: 72815713