Handle task cancellation gracefully in SDK manager installation ops.
This CL ensures graceful handling of the case when SDK
packages installation is being cancelled via IDEA UI.

Before, the IDEA Task's progress indicator was not checked
with checkCancelled(), which led to significant latency between pressing
Cancel and getting the task actually cancelled. Orthogonally to that,
there could also be random IO exceptions caused by installer
operation being invoked on files which were not downloaded
due to cancellation, but the control flow proceeded because the
ProcessCancelledException was always swallowed.

Test: Adjusted InstallTaskTest
Bug: b/72806377
Change-Id: Ic91bec7f06e24051978d299d5551f28fc5281d76