Designer tests clean up mock KeyboardFocusManager
... otherwise all sorts of memory leaks happen (see bug for more
details) because of Mockito's ongoing stubbing.

We can make EditorFixtureBase a Closeable resource, so that when it
is not escaping from the current method it can be enclosed by
try-with-resources in Java and use in Kotlin, thus minimizing the
chance of forgetting to call tearDown or bailing out early due to an
uncaught exception.

Test: these
Bug: 72755830
Change-Id: I0e298acb2fcd70589c75d8c307555c98a76d175b