What's New Assistant
To edit, change android/resources/bundle/whatsNew/3.2.xml
Have to rebuild AS to see the change. Looks like resources gets packaged
in the build process.

Current behaviour: Opens the Whats New assistants only if:
1. The feature is enabled (false by default)
2. There is a xml file checked that matches the current studio version

So if we don't get around to adding an xml file then this feature has no
effect on AS.

Bug: b/72815475
Test: WhatsNewAssistantBundleCreatorTest
Change-Id: I76c6ee345e508773e25794476f54b18adfa69b47