Remove Wait that doesn't fix any problem
This Wait did not fix the original problem it was meant to solve. The
original problem was to wait for the cursor to move to the place that we
expected it to click on.  However, there is no guarantee the cursor will
move there, such as when IntelliJ is collapsing single-line functions,
or collapsing some lines of anonymous classes. In some cases, the cursor
didn't move at all.

Test: existing tests
Change-Id: I4aa93dc64b6126a5e342eae28c22c51d37e08e99