Update all JUnit BatchInvoker tests to use BatchInvokerStrategyRule
This rule, added recently, is a bit less verbose and safer to use than
its setUp/tearDown counterpart. While the code that was changed here
was already working fine as is, this at least ensures any future tests,
which may end up forking existing tests, are more likely to use the
rule.

Test: Updated
Change-Id: I3cdc3b552f0864ba88fdc409d8ca88d8630cf444