Extract build system specific runner logic out of GuiTestRunner
I75c56ef97 introduced a bunch of changes to GuiTestRunner in order to
support running the same UI test for multiple build systems.
I8c6f3b562 further customized GuiTestRunner to include the build system
name as part of the test name.

However, the second addition introduced a regression: For tests that are
not run with the MultiBuildGuiTestRunner, the new name was still being
used. This resulted in the IntelliJ UI displaying the build system name
as opposed to the test class name for such tests.

This CL fixes the issue by moving all logic specific to supporting the
running of UI tests GuiTestRunner out into a subclass that is only used
by the MultiBuildGuiTestRunner.

Fixes: 72887718
Test: existing UI tests
Change-Id: Ia44537f85949dfc3bbe2761c11c950fedb50cbdd