Don't create empty build system specific test suites
MultiBuildGuiTestRunner always created a runner per build system even if
a class didn't have any tests that were marked to run with one of those
build systems. This results in a Suite with no tests, which breaks
expectations in TestGroupFilter.

This CL fixes MultiBuildGuiTestRunner such that it only creates runners
for a particular build system only when there are tests that expect to
be run with that target build system.

Test: existing UI tests
Fixes: 72887846
Change-Id: I80add5dc0008f7c765dba948cb6b1500696ef729