Expand the timeouts for createAndRunInstantApp.
This test goes through the new project wizard flow. When setting up a
new project, the IDE actually downloads a Gradle distribution from
services.gradle.org rather than retrieving the distribution from the
local filesystem. This makes the test very slow and its success depends
on the network.

A waitForBackgroundTasks was also removed because it led to a race
condition. The wait was initially meant to wait for the first progress
bar to disappear, but the wait can actually end up seeing the second
progress bar. This causes the second set of progress bars to not have a
wait that lasted long enough.

Bug: 72832198
Test: createAndRunInstantApp
Change-Id: I2de608725b569fefbbd9af88e28c9ec520e6ea5e