Export CPU trace files using a contextual menu
Adds a simple contextual/popup menu to CPU profiler L2 containing the
actions common to all the profilers and add another option to export
trace files using a file picker dialog.

This feature is added under a flag enabled by default. In case we
identify an issue in the future, we should disable it.

Change-Id: I5b3b83733ae7b1f8209294ab61c075156112a6b3
Fixes: 72982838
Bug: 72982718
Bug: 33466955
Test: CpuProfilerStageViewTest#contextMenuShouldBeInstalled,
      CpuProfilerStageTest#traceFilesGeneratedPerTrace,
      CpuProfilerStageTest#rangeIntersectionReturnsASingleTraceId.
      Also, context menu can be manually tested by opening CPU profiler L2
      and right-clicking a capture ("Export trace..." should be enabled)
      or an area outside a capture ("Export trace..." should be disabled).
      After exporting a trace file, check it was saved to the right folder.