Ensure Bazel UI tests wait for sync to start
The Bazel-specific UI tests in NlEditorTest fail sporadically
due to a race condition when waiting for project syncs to finish.
The tests determine whether or not a sync has finished by waiting
for all background tasks to complete. In the problematic case,
the wait for background tasks to finish happens before the sync
starts, causing the test to continue without actually waiting
on the sync.

This CL resolves the issue by ensuring that Bazel UI tests wait
for project sync to start before checking the progress of
background tasks. This is accomplished by having
BazelGuiTestProjectSystem#waitForProjectSyncToFinish first wait
until the Bazel console reports that a sync has started. To make
sure that the message in the Bazel console corresponds to the
desired sync and not a previous one, BazelGuiTestProjectSystem
now clears the Bazel console before initiating a sync.

Logging statements have been added to the Bazel sync workflow
to help future debugging.

Test: Ran NlEditorTest#designEditorUnavailableIf* 100 times
      with no flakes. Previously we would see a failure for
      these tests about 1 out of 10 runs.
Change-Id: I822cc1f48bddd06a9d67ef3d6d95c896c551224d
Fixes: 72912068
Test-Strategy: ui.test.groups=DEFAULT