Fix memory leak in ActionsToolbar
ActionsToolbar registers itself as a listener on the Configuration
object of its DesignSurface. The configuration comes from the NlModel
and DesignSurface may change its model during its lifetime. Therefore,
configuration object in ActionsToolbar.getToolbarComponent() method
(where the listener is registered) may be different than the
configuration object in dispose() method and in such cases
removeListener() call does nothing, resulting in a memory leak.

The fix adresses the memory leak (it makes sure it is the same
Configuration object that receives both addListener and removeListener
calls).

Other things addressed in the fix:
* Removed unnecessary myModel field
* Changing model on a DesignSurface will now move the
  ActionsToolbar's listener to the Configuration object associated
  with the new model.

Test: N/A
Change-Id: I4bd4f4049ca0e45f2d4955aa01dffe33ec33c549
Fixes: 73080536