[New PSD] Collection editor: store ParsedValue in table model.
When editing maps it should be possible to edit the value of a map entry
first or change the entry key by removing the old one and later provinding
a new one. This scenario requires a temporary storage for the map values.

The CL changes what is stored in the table model from plain strings to
wrapped ParsedValues.

Bug: 72738791
Test: UI tests will be added when the PSD is completed
Change-Id: Id5c8e87253dcf0f66cd0611492bf6d92138a6212