Resolve classpath conflict in Bazel SimpleLocalApplication
Currently, the Bazel BUILD file for SimpleLocalApplication introduces
a classpath conflict by specifying the same custom_package in the
android_binary rule and the "activities" android_library rule. This
resulted in a "Multiple R classes generated with the same java
package" warning spamming stdio every time the project was synced
during a UI test.

This CL resolves the issue by changing the custom_package attribute
of the android_binary to "google" instead of
"google.simpleapplication", in keeping with the build tutorial on
docs.bazel.build.

Fixes: 73115571
Test: Ran NlEditorTest#designEditorUnavailableIfLastBazelSyncFailed
      and confirmed that the warning message is no longer being
      logged.

Change-Id: Id9232abd8e9bab62618803d7b35079e6238a6d17