Clean up export dialog
This CL remove the ExportDialog so that user can export the file
in one step. This CL also provide a default file name.
Test steps
1. Open memory profiler and click "dump java heap" button
2. After the heap dump ends, click "Export capture to file"
3. Before this CL, you will see a save dialog that only allows
you to type the path direcly and you need to click another
button to enter the Menu where path can be selected interactively.
4. Now, You should see a interactive Save Dialog popped up
with "heapdump" as the default file name
5. Click "Save" after selecting the path and you should find
the hprof file  in the selected location.

Bug: 67717673
Test: Manual test
Change-Id: Id49ddbdc6ee62141c4e479effdca127377580f35