Check for bazel plugin in bazel ui-tests.
When a bazel project uitest runs in an environment where the bazel
plugin is not present, the test will fail with an error saying the
fest bot cannot find a button on the welcome frame. This does not
address the true cause of the problem and does not help the tester
resolve the issue.

This CL adds the method validateSetup() in GuiTestProjectSystem
to ensure the test is setup correctly.  This method is called in
CurrentGuiTestProjectSystem#before() before continuing with test
operations.

If the bazel plugin is missing then the test will output an
IllegalStateException with information detailing on how to resolve the
issue.

Test: Ran bazel uitests with plugin missing and verified the message
shows correctly. (and vice versa)
Bug: 73254021

Change-Id: I70277fcca25d2f9675bae7f167002effe4794c15