Add Record action to CPU profiler context menu
Added the option to start/stop CPU captures using the context menu
triggered when right-clicking the CPU usage chart.

That CL required a small refactoring in ProfilerAction class to add the
ability to have multiple icon and text for the same button. It adds
ProfilerActionTest to test the new (and old) behaviors of ProfilerAction.

Bug: 72982718
Test: CpuProfilerStageViewTest#contextMenuShouldBeInstalled and
      Added ProfilerActionTest.

Change-Id: Icd8675fe8454d8c4e9d24489c9dc023633119663