Gracefully handle null target in ThemeResolver
I saw this throw an AssertionError when getConfiguredFrameworkResources
with a null target. I cannot relibaly reproduce, but getTarget() is
indeed nullable, so we should be prepared. Returning an empty list is
what the code used to do before switching to the new repo
implementation.

Test: I don't know a way to reproduce it, probably not worth the effort
of getting the project into a weird state between syncs when available
API levels are not known.

Change-Id: I2ee5cb1a48085436e966b1e14760b24ef22dc321