Include attachments in GuiTests error message
The UI test suite occasionally throws errors from the
highlighting pass.  However, the error message does not include
vital information such as the original stack trace or the name of
the source file that the highlighter failed on.

Here's how the highting	infrastructure deals with an exception:

  LOG.error("annotator: " + annotator +	...,
     new Attachment("root_path.txt", file.getPath()),
     new Attachment("stack", t));

The actual throwable (t) isn't included	in the error, but it's
attaching vital	information, including the stack, as attachments.

This CL updates the GuiTests fixture to look up the error message
attachments and to include the string representation of these in
the error message, to help track down this issue.

Bug: 72880366
Change-Id: I991ac9196c99976a638e61d271a8c6d4a47a0c3c